WIP: first cut at an alternate version of env_utils for an orchestrated system.
