+++ test/test_s3_utils.py
@@ -5,7 +5,7 @@ import pytest
 
 from dcicutils import s3_utils as s3_utils_module
 from dcicutils.beanstalk_utils import compute_ff_prd_env, compute_cgap_prd_env, compute_cgap_stg_env
-from dcicutils.env_utils import get_standard_mirror_env, FF_PUBLIC_URL_STG, FF_PUBLIC_URL_PRD, CGAP_PUBLIC_URL_PRD
+from dcicutils.env_utils_legacy import get_standard_mirror_env, FF_PUBLIC_URL_STG, FF_PUBLIC_URL_PRD, CGAP_PUBLIC_URL_PRD
 from dcicutils.exceptions import SynonymousEnvironmentVariablesMismatched
 from dcicutils.misc_utils import ignored, ignorable
 from dcicutils.qa_utils import override_environ, MockBoto3
